requirements.txt
streamlit
torch
torchvision
opencv-python-headless
Pillow
huggingface_hub